fix happ-entrypoint.sh if分岐追加

+++ hyperapp/happ-entrypoint.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/sh
-sudo mv /var/www/happ-pre/* /var/www/happ/
-sudo mv /var/www/happ-pre/.??* /var/www/happ/
-sudo rmdir /var/www/happ-pre
+if [ -f /var/www/happ/package.json ]; then
+    sudo npm install
+    sudo npm run build
+else
+    sudo mv /var/www/happ-pre/* /var/www/happ/
+    sudo mv /var/www/happ-pre/.??* /var/www/happ/
+    sudo rmdir /var/www/happ-pre
+fi
 sudo chown -R 1000:1000 /var/www/happ
 sudo nginx -g 'daemon off;'
